refactor: Removed ObservedSynchrosNodes
